requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.6
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
phonenumbers==8.13.7
psycopg2-binary==2.9.5
sqlparse==0.4.3
